split BulkCreateUserCsvFile from base.py
